refactor(pipeline): delete the dead legacy in-process pipeline
The pre-Temporal orchestrator (run_discovery_legacy_once), its stage-runner
map, _build_stage_kwargs, _resolve_stages, and the discovery enrichment
worker thread helpers had no production callers - the Temporal DiscoverWorkflow
path replaced them (2026-07-03 ADRs). They survived only through tests that
exercised the dead entrypoint itself.

- delete the legacy orchestrator subtree from pipeline/runner.py (~560 lines);
  the live per-stage runners, drain machinery, and _run_stage_observed remain
- keep the _StageRunner alias with its one live consumer, _run_stage_observed
- stop re-exporting _PENDING_SQL/_count_pending/_resolve_stages as package
  API; tests import the module directly
- delete the dead-path tests and the autouse fixture that stubbed the
  removed worker helpers; live-path tests (score activity span, _run_tailor
  contract, smart-extract, _run_enrich propagation) stay
