+++ workers/automation/src/jobctrl/pipeline/__init__.py
@@ -10,9 +10,6 @@ from jobctrl.pipeline.runner import (
     STAGE_META,
     STAGE_ORDER,
     SUPPORTED_STAGE_ORDER,
-    _PENDING_SQL,
-    _count_pending,
-    _resolve_stages,
     run_single_job,
 )
 
@@ -22,8 +19,5 @@ __all__ = [
     "STAGE_META",
     "STAGE_ORDER",
     "SUPPORTED_STAGE_ORDER",
-    "_PENDING_SQL",
-    "_count_pending",
-    "_resolve_stages",
     "run_single_job",
 ]
