+++ workers/automation/src/jobctrl/pipeline/runner.py
@@ -322,6 +322,11 @@ def _stage_counts(result: dict[str, Any] | None) -> dict[str, Any]:
     return result
 
 
+# A stage runner is any callable returning the stage result dict. ``Callable``
+# (lowercase ``callable`` is the runtime predicate, not a generic type alias).
+_StageRunner = Callable[..., dict[str, Any]]
+
+
 def _run_stage_observed(
     stage: str,
     runner: _StageRunner,
@@ -1969,443 +1974,6 @@ def _jobspy_progress_callback(
 # Individual stage runners
 # ---------------------------------------------------------------------------
 
-def run_discovery_legacy_once(
-    workers: int = 1,
-    limit: int = 0,
-    min_score: int = 7,
-    validation_mode: str = "normal",
-    llm_model: str | None = DEFAULT_PIPELINE_LLM_MODEL_SPEC,
-    tailor_models: tuple[str, ...] = (),
-    tailor_judge_model: str | None = None,
-    tailor_judge_min_score: float | None = None,
-    source_ids: tuple[str, ...] = (),
-    cancel_event: threading.Event | None = None,
-) -> dict:
-    """Stage: broad-board, Workday, and smart-extract discovery."""
-    stats: dict = {"jobspy": None, "workday": None, "smartextract": None}
-    source_results: dict[str, Any] = {}
-    source_failures: dict[str, str] = {}
-    source_succeeded = False
-    selected_source_ids = tuple(dict.fromkeys(source_id.strip() for source_id in source_ids if source_id.strip()))
-    source_filter_active = bool(selected_source_ids)
-    provided_cancel_event = cancel_event
-    cancel_event = cancel_event or threading.Event()
-    conn = init_db()
-    try:
-        seed_discovery_control_queues(conn, config.load_source_registry())
-    except Exception:
-        log.debug("Failed to seed discovery control queues", exc_info=True)
-    schedule = _plan_discovery_schedule(limit, source_ids=selected_source_ids)
-    bounded_workers = max(1, workers)
-    start_count = _pipeline_job_count() if limit > 0 else 0
-    jobspy_sources = schedule.for_prefix("jobspy")
-    ats_sources = tuple(
-        source
-        for source in schedule.for_kinds(SourceKind.ATS_API)
-        if not source.source_id.startswith("workday:")
-    )
-    workday_sources = schedule.for_prefix("workday")
-    smart_extract_sources = _smart_extract_sources(schedule)
-    source_step_count = (
-        (1 if (not source_filter_active or jobspy_sources) else 0)
-        + (1 if ats_sources else 0)
-        + (1 if (not source_filter_active or workday_sources) else 0)
-        + (1 if (not source_filter_active or smart_extract_sources) else 0)
-    )
-    progress_total = source_step_count + 2
-    progress_completed = 0
-
-    # Broad-board JobStreaming — skip if disabled or unavailable.
-    search_cfg = config.load_search_config() or {}
-    def run_hygiene(label: str) -> int:
-        hygiene = retire_invalid_source_jobs(
-            conn,
-            search_cfg=search_cfg,
-            run_id=f"discovery:hygiene:{label}",
-        )
-        retired = int(hygiene.get("retired_jobs") or 0)
-        if retired:
-            console.print(f"  [yellow]Discovery hygiene retired {retired} invalid source jobs[/yellow]")
-        return retired
-
-    try:
-        run_hygiene("before")
-    except Exception:
-        log.warning("Discovery hygiene failed", exc_info=True)
-    enrichment_kwargs: dict[str, Any] = {"workers": bounded_workers, "limit": limit}
-    if provided_cancel_event is not None:
-        enrichment_kwargs["cancel_event"] = cancel_event
-    enrichment_done, enrichment_result, enrichment_thread = _start_discovery_enrichment_worker(**enrichment_kwargs)
-
-    def finish_discovery() -> dict:
-        nonlocal progress_completed
-        _record_pipeline_event(
-            "discover",
-            "StageStarted",
-            "info",
-            "Discovery detail enrichment finishing",
-            _discovery_progress_payload(
-                completed=progress_completed,
-                total=progress_total,
-                current_step="Detail enrichment",
-                message="Detail enrichment finishing",
-            ),
-        )
-        enrichment_stats = _finish_discovery_enrichment_worker(
-            enrichment_done,
-            enrichment_thread,
-            enrichment_result,
-        )
-        stats["enrichment"] = str(enrichment_stats.get("status", "ok"))
-        progress_completed += 1
-        _record_pipeline_event(
-            "discover",
-            "StageCompleted",
-            "info",
-            "Discovery detail enrichment complete",
-            _discovery_progress_payload(
-                completed=progress_completed,
-                total=progress_total,
-                current_step="Detail enrichment",
-                message="Detail enrichment complete",
-            ),
-        )
-        try:
-            run_hygiene("after")
-        except Exception:
-            log.warning("Post-discovery hygiene failed", exc_info=True)
-        _record_pipeline_event(
-            "discover",
-            "StageStarted",
-            "info",
-            "Discovery preparation started",
-            _discovery_progress_payload(
-                completed=progress_completed,
-                total=progress_total,
-                current_step="Preparation",
-                message="Preparation started",
-            ),
-        )
-        try:
-            from jobctrl.pipeline.preparation import start_discovery_preparation_workflows
-
-            preparation_stats = start_discovery_preparation_workflows(
-                min_score=min_score,
-                limit=limit,
-                workers=bounded_workers,
-                validation_mode=validation_mode,
-                llm_model=llm_model,
-                tailor_models=tailor_models,
-                tailor_judge_model=tailor_judge_model,
-                tailor_judge_min_score=tailor_judge_min_score,
-            )
-            if preparation_stats.get("has_work"):
-                stats["preparation"] = str(preparation_stats.get("status", "ok"))
-                stats["preparation_counts"] = preparation_stats
-                if preparation_stats.get("status") != "ok":
-                    stats["status"] = "partial"
-            preparation_status = "partial" if preparation_stats.get("status") not in (None, "ok") else "succeeded"
-            progress_completed += 1
-            _record_pipeline_event(
-                "discover",
-                "StageCompleted",
-                "warn" if preparation_status == "partial" else "info",
-                "Discovery preparation complete",
-                _discovery_progress_payload(
-                    completed=progress_completed,
-                    total=progress_total,
-                    current_step="Preparation",
-                    status=preparation_status,
-                    message="Preparation complete",
-                ),
-            )
-        except Exception as exc:
-            log.exception("Discovery preparation orchestration failed")
-            stats["preparation"] = "failed"
-            stats["preparation_error"] = str(exc)
-            stats["status"] = "partial"
-            progress_completed += 1
-            _record_pipeline_event(
-                "discover",
-                "StageFailed",
-                "error",
-                f"Discovery preparation failed: {exc}",
-                _discovery_progress_payload(
-                    completed=progress_completed,
-                    total=progress_total,
-                    current_step="Preparation",
-                    status="failed",
-                    message="Preparation failed",
-                ),
-            )
-        return stats
-
-    def _run_source_isolated(
-        key: str,
-        label: str,
-        sources: tuple[ScheduledSource, ...],
-        run_fn: Callable[..., Any],
-    ) -> None:
-        """Run one discovery source group, tolerating a per-source failure.
-
-        A raised ``SourceUnavailableError`` has already been durably recorded
-        inside ``_run_discovery_source`` (DiscoveryRunFailed event, StageFailed
-        pipeline event, failed operational attempt). Catch it here so the
-        remaining source groups still run and the stage yields the healthy
-        sources' jobs as a partial result. When every source fails the caller
-        raises an aggregated ``SourceUnavailableError`` from the stage.
-        """
-        nonlocal progress_completed, source_succeeded
-        try:
-            status = _run_discovery_source(
-                key,
-                label,
-                sources,
-                run_fn,
-                progress_completed=progress_completed,
-                progress_total=progress_total,
-            )
-            stats[key] = status
-            if status == "ok":
-                source_succeeded = True
-        except SourceUnavailableError as exc:
-            cause = str(exc)
-            source_failures[label] = cause
-            stats[key] = "failed"
-            stats[f"{key}_error"] = cause
-            stats["status"] = "partial"
-            console.print(f"  [red]{label} error:[/red] {cause}")
-        progress_completed += 1
-
-    def run_jobspy(run_id: str | None = None) -> dict:
-        if search_cfg.get("disable_jobspy", False):
-            console.print("  [dim]Broad-board discovery disabled in settings[/dim]")
-            result = {"new": 0, "existing": 0, "errors": 0, "db_total": 0, "queries": 0}
-            source_results["jobspy"] = result
-            return result
-        console.print("  [cyan]Broad-board JobStreaming crawl...[/cyan]")
-        try:
-            from jobctrl.discovery.jobspy import run_discovery
-        except ImportError:
-            console.print("  [dim]JobStreaming not installed — skipping[/dim]")
-            result = {"new": 0, "existing": 0, "errors": 0, "db_total": 0, "queries": 0}
-            source_results["jobspy"] = result
-            return result
-        jobspy_cfg = _jobspy_config_for_sources(search_cfg, jobspy_sources)
-        if not jobspy_cfg.get("boards"):
-            console.print("  [dim]No runnable broad boards scheduled[/dim]")
-            result = {"new": 0, "existing": 0, "errors": 0, "db_total": 0, "queries": 0}
-            source_results["jobspy"] = result
-            return result
-        jobspy_limit = _scheduled_limit(schedule, "jobspy", limit)
-        try:
-            signature = inspect.signature(run_discovery)
-        except (TypeError, ValueError):
-            accepts_run_id = False
-            accepts_progress_callback = False
-            accepts_cancel_event = False
-        else:
-            accepts_run_id = (
-                "run_id" in signature.parameters
-                or any(
-                    param.kind is inspect.Parameter.VAR_KEYWORD
-                    for param in signature.parameters.values()
-                )
-            )
-            accepts_progress_callback = (
-                "progress_callback" in signature.parameters
-                or any(
-                    param.kind is inspect.Parameter.VAR_KEYWORD
-                    for param in signature.parameters.values()
-                )
-            )
-            accepts_cancel_event = (
-                "cancel_event" in signature.parameters
-                or any(
-                    param.kind is inspect.Parameter.VAR_KEYWORD
-                    for param in signature.parameters.values()
-                )
-            )
-
-        run_kwargs: dict[str, Any] = {"cfg": jobspy_cfg, "limit": jobspy_limit}
-        if accepts_run_id:
-            run_kwargs["run_id"] = run_id
-        if accepts_progress_callback and run_id:
-            def record_jobspy_progress(snapshot: dict[str, Any]) -> None:
-                error_count = _first_present(
-                    snapshot.get("errors"),
-                    snapshot.get("error_count"),
-                    snapshot.get("errorCount"),
-                )
-                _record_discovery_source_progress(
-                    source="jobspy",
-                    label="Broad boards",
-                    run_id=run_id,
-                    source_ids=tuple(item.source_id for item in jobspy_sources if item.should_run),
-                    progress_completed=progress_completed,
-                    progress_total=progress_total,
-                    source_progress=DiscoveryRunProgress(
-                        completed=int(snapshot.get("completed") or 0),
-                        total=int(snapshot.get("total") or 0),
-                        unit=str(snapshot.get("unit") or "searches"),
-                        current_query=_optional_text(
-                            _first_present(snapshot.get("current_query"), snapshot.get("currentQuery"))
-                        ),
-                        current_location=_optional_text(
-                            _first_present(snapshot.get("current_location"), snapshot.get("currentLocation"))
-                        ),
-                        new_jobs=_optional_int(_first_present(snapshot.get("new_jobs"), snapshot.get("newJobs"))),
-                        existing_jobs=_optional_int(
-                            _first_present(snapshot.get("existing_jobs"), snapshot.get("existingJobs"))
-                        ),
-                        filtered_jobs=_optional_int(
-                            _first_present(snapshot.get("filtered_jobs"), snapshot.get("filteredJobs"))
-                        ),
-                        error_count=_optional_int(error_count),
-                        raw_total=_optional_int(_first_present(snapshot.get("raw_total"), snapshot.get("rawTotal"))),
-                        recovered_units=_optional_int(
-                            _first_present(
-                                snapshot.get("recovered_units"),
-                                snapshot.get("recoveredUnits"),
-                            )
-                        ),
-                    ),
-                    message=str(snapshot.get("message") or "JobStreaming progress updated"),
-                )
-
-            run_kwargs["progress_callback"] = record_jobspy_progress
-        if accepts_cancel_event and provided_cancel_event is not None:
-            run_kwargs["cancel_event"] = cancel_event
-        source_results["jobspy"] = run_discovery(**run_kwargs)
-        return source_results["jobspy"]
-
-    if not source_filter_active or jobspy_sources:
-        _run_source_isolated("jobspy", "Broad boards", jobspy_sources, run_jobspy)
-    if _discover_limit_consumed(start_count, limit, source_results.get("jobspy")):
-        if ats_sources:
-            stats["ats_api"] = _skip_discovery_source(
-                "ats_api",
-                "Canonical ATS APIs",
-                "limit reached",
-                progress_completed=progress_completed,
-                progress_total=progress_total,
-            )
-            progress_completed += 1
-        if not source_filter_active or workday_sources:
-            stats["workday"] = _skip_discovery_source(
-                "workday",
-                "Workday scraper",
-                "limit reached",
-                progress_completed=progress_completed,
-                progress_total=progress_total,
-            )
-            progress_completed += 1
-        if not source_filter_active or smart_extract_sources:
-            stats["smartextract"] = _skip_discovery_source(
-                "smartextract",
-                "Smart extract",
-                "limit reached",
-                progress_completed=progress_completed,
-                progress_total=progress_total,
-            )
-            progress_completed += 1
-        return finish_discovery()
-
-    if ats_sources:
-        def run_ats(run_id: str | None = None) -> dict:
-            console.print("  [cyan]Canonical ATS APIs...[/cyan]")
-            ats_kwargs: dict[str, Any] = {
-                "search_cfg": search_cfg,
-                "run_id": run_id or f"discovery:ats_api:{uuid.uuid4().hex}",
-                "limit": _scheduled_limit_for_sources(ats_sources, _discover_remaining_limit(start_count, limit)),
-            }
-            if provided_cancel_event is not None:
-                ats_kwargs["cancel_event"] = cancel_event
-            source_results["ats_api"] = run_scheduled_ats_sources(conn, ats_sources, **ats_kwargs)
-            return source_results["ats_api"]
-
-        _run_source_isolated("ats_api", "Canonical ATS APIs", ats_sources, run_ats)
-        if _discover_limit_consumed(start_count, limit, source_results.get("ats_api")):
-            if not source_filter_active or workday_sources:
-                stats["workday"] = _skip_discovery_source(
-                    "workday",
-                    "Workday scraper",
-                    "limit reached",
-                    progress_completed=progress_completed,
-                    progress_total=progress_total,
-                )
-                progress_completed += 1
-            if not source_filter_active or smart_extract_sources:
-                stats["smartextract"] = _skip_discovery_source(
-                    "smartextract",
-                    "Smart extract",
-                    "limit reached",
-                    progress_completed=progress_completed,
-                    progress_total=progress_total,
-                )
-                progress_completed += 1
-            return finish_discovery()
-
-    # Workday corporate scraper
-    def run_workday(run_id: str | None = None) -> dict:
-        console.print("  [cyan]Workday corporate scraper...[/cyan]")
-        from jobctrl.discovery.workday import run_workday_discovery
-        workday_kwargs: dict[str, Any] = {
-            "employers": _workday_employers_for_sources(workday_sources),
-            "workers": bounded_workers,
-            "limit": _scheduled_limit(schedule, "workday", _discover_remaining_limit(start_count, limit)),
-            "run_id": run_id,
-        }
-        if provided_cancel_event is not None:
-            workday_kwargs["cancel_event"] = cancel_event
-        source_results["workday"] = run_workday_discovery(**workday_kwargs)
-        return source_results["workday"]
-
-    if not source_filter_active or workday_sources:
-        _run_source_isolated("workday", "Workday scraper", workday_sources, run_workday)
-        if _discover_limit_consumed(start_count, limit, source_results.get("workday")):
-            if not source_filter_active or smart_extract_sources:
-                stats["smartextract"] = _skip_discovery_source(
-                    "smartextract",
-                    "Smart extract",
-                    "limit reached",
-                    progress_completed=progress_completed,
-                    progress_total=progress_total,
-                )
-                progress_completed += 1
-            return finish_discovery()
-
-    # Smart extract
-    def run_smart_extract_source() -> dict:
-        console.print("  [cyan]Smart extract (AI-powered scraping)...[/cyan]")
-        enqueue_manual_action_for_sources(conn, smart_extract_sources)
-        from jobctrl.discovery.smartextract import run_smart_extract
-        smart_kwargs: dict[str, Any] = {
-            "sites": _smart_extract_sites(smart_extract_sources),
-            "workers": bounded_workers,
-            "limit": _scheduled_limit_for_sources(
-                smart_extract_sources,
-                _discover_remaining_limit(start_count, limit),
-            ),
-        }
-        if provided_cancel_event is not None:
-            smart_kwargs["cancel_event"] = cancel_event
-        source_results["smartextract"] = run_smart_extract(**smart_kwargs)
-        return source_results["smartextract"]
-
-    if not source_filter_active or smart_extract_sources:
-        _run_source_isolated(
-            "smartextract", "Smart extract", smart_extract_sources, run_smart_extract_source
-        )
-
-    result = finish_discovery()
-    if source_failures and not source_succeeded:
-        raise SourceUnavailableError(
-            "All discovery sources failed: " + "; ".join(source_failures.values())
-        )
-    return result
-
-
 def _run_enrich(
     workers: int = 1,
     limit: int = 0,
@@ -2576,94 +2144,6 @@ def _run_cover(
         "skipped": skipped,
         "selected": len(jobs),
     }
-
-
-# Map stage names to their runner functions. ``Callable`` (lowercase
-# ``callable`` is the runtime predicate, not a generic type alias).
-_StageRunner = Callable[..., dict[str, Any]]
-_STAGE_RUNNERS: dict[str, _StageRunner] = {
-    "discover": run_discovery_legacy_once,
-    "enrich":   _run_enrich,
-    "score":    _run_score,
-    "tailor":   _run_tailor,
-    "cover":    _run_cover,
-}
-
-
-def _build_stage_kwargs(
-    stage: str,
-    *,
-    min_score: int = 7,
-    workers: int = 1,
-    validation_mode: str = "normal",
-    limit: int = 0,
-    rescore: bool = False,
-    retailor: bool = False,
-    llm_model: str | None = DEFAULT_PIPELINE_LLM_MODEL_SPEC,
-    tailor_models: tuple[str, ...] = (),
-    tailor_judge_model: str | None = None,
-    tailor_judge_min_score: float | None = None,
-    source_ids: tuple[str, ...] = (),
-    cancel_event: threading.Event | None = None,
-) -> dict:
-    """Build the keyword arguments for a stage runner."""
-    kwargs: dict = {}
-
-    if stage in ("discover", "enrich", "score"):
-        kwargs["workers"] = workers
-    if stage in ("discover", "enrich", "score"):
-        kwargs["limit"] = limit
-    if cancel_event is not None and stage in ("discover", "enrich", "score", "tailor", "cover"):
-        kwargs["cancel_event"] = cancel_event
-    if stage == "discover":
-        kwargs["min_score"] = min_score
-        kwargs["validation_mode"] = validation_mode
-        kwargs["llm_model"] = llm_model
-        kwargs["tailor_models"] = tailor_models
-        kwargs["tailor_judge_model"] = tailor_judge_model
-        kwargs["tailor_judge_min_score"] = tailor_judge_min_score
-        if source_ids:
-            kwargs["source_ids"] = source_ids
-    elif stage == "score":
-        kwargs["rescore"] = rescore
-        kwargs["llm_model"] = llm_model
-    elif stage in ("tailor", "cover"):
-        kwargs["min_score"] = min_score
-        kwargs["limit"] = limit
-        kwargs["validation_mode"] = validation_mode
-        kwargs["llm_model"] = llm_model
-        if stage == "tailor":
-            kwargs["workers"] = workers
-            kwargs["retailor"] = retailor
-            kwargs["tailor_models"] = tailor_models
-            kwargs["tailor_judge_model"] = tailor_judge_model
-            kwargs["tailor_judge_min_score"] = tailor_judge_min_score
-
-    return kwargs
-
-
-# ---------------------------------------------------------------------------
-# Stage resolution
-# ---------------------------------------------------------------------------
-
-def _resolve_stages(stage_names: list[str]) -> list[str]:
-    """Resolve 'all' and validate/order stage names."""
-    if "all" in stage_names:
-        return list(PRIMARY_STAGE_ORDER)
-
-    resolved = []
-    for name in stage_names:
-        if name not in STAGE_META:
-            console.print(
-                f"[red]Unknown stage:[/red] '{name}'. "
-                f"Available: {', '.join((*SUPPORTED_STAGE_ORDER, 'enrich'))}, all"
-            )
-            raise SystemExit(1)
-        if name not in resolved:
-            resolved.append(name)
-
-    # Maintain canonical order
-    return [s for s in INTERNAL_STAGE_ORDER if s in resolved]
 
 
 # SQL to count pending work for each stage. Round-1 review B1: every
@@ -3235,43 +2715,6 @@ def _run_discovery_enrichment_until_idle(
         )
 
 
-def _start_discovery_enrichment_worker(
-    *,
-    workers: int,
-    limit: int,
-    cancel_event: threading.Event | None = None,
-) -> tuple[threading.Event, dict[str, Any], threading.Thread]:
-    """Start Discovery's internal detail-enrichment queue worker."""
-    discovery_done = threading.Event()
-    result: dict[str, Any] = {}
-    thread = threading.Thread(
-        target=_run_discovery_enrichment_until_idle,
-        kwargs={
-            "discovery_done": discovery_done,
-            "result": result,
-            "workers": workers,
-            "limit": limit,
-            "cancel_event": cancel_event,
-        },
-        name="discover-detail-enrichment",
-        daemon=True,
-    )
-    thread.start()
-    return discovery_done, result, thread
-
-
-def _finish_discovery_enrichment_worker(
-    discovery_done: threading.Event,
-    thread: threading.Thread,
-    result: dict[str, Any],
-) -> dict[str, Any]:
-    """Signal Discovery's internal detail-enrichment queue worker and wait for it."""
-    discovery_done.set()
-    thread.join()
-    return result or {"status": "ok", "passes": 0, "pending": 0}
-
-
-# ---------------------------------------------------------------------------
 # Single-job processing
 # ---------------------------------------------------------------------------
